requirements.txt
openai
watchdog
python-dotenv
